fragment_tools.xml --renamed--> fragment_setting.xml also updated: android:id changed: @+id/text_tools --changed--> @+id/text_setting
* ui/tools package --renamed--> ui/setting
* ToolsFragment --renamed--> SettingFragment

* ToolsViewModel --renamed--> SettingViewModel

strings.xml updated:
nav_tools --changed--> menu_video and it's label changed to "Videos"

activity_main_drawer.xml updated:
android:id changed: @+id/nav_tools --changed--> @+id/nav_setting
android:title changed: @string/menu_tools --changed--> @string/menu_setting

mobile_navigation.xml updated:
android:id changed: @+id/nav_tools --> @+id/nav_setting
android:name changed
android:label changed: @string/menu_tools --> @string/menu_setting
tools:layout changed @layout/fragment_tools --> @layout/fragment_setting
