+++ app/src/main/java/com/fanaye/sorantravel/ui/setting/SettingViewModel.java
@@ -1,16 +1,16 @@
-package com.fanaye.sorantravel.ui.tools;
+package com.fanaye.sorantravel.ui.setting;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ToolsViewModel extends ViewModel {
+public class SettingViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public ToolsViewModel() {
+    public SettingViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is tools fragment");
+        mText.setValue("This is setting fragment");
     }
 
     public LiveData<String> getText() {
