+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -29,9 +29,9 @@
             android:icon="@drawable/ic_menu_where_to_go"
             android:title="@string/menu_where_to_go" />
         <item
-            android:id="@+id/nav_tools"
+            android:id="@+id/nav_setting"
             android:icon="@drawable/ic_menu_setting"
-            android:title="@string/menu_tools" />
+            android:title="@string/menu_setting" />
     </group>
 
     <item android:title="Communicate">
