+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -34,10 +34,10 @@
         tools:layout="@layout/fragment_slideshow" />
 
     <fragment
-        android:id="@+id/nav_tools"
-        android:name="com.fanaye.sorantravel.ui.tools.ToolsFragment"
-        android:label="@string/menu_tools"
-        tools:layout="@layout/fragment_tools" />
+        android:id="@+id/nav_setting"
+        android:name="com.fanaye.sorantravel.ui.setting.SettingFragment"
+        android:label="@string/menu_setting"
+        tools:layout="@layout/fragment_setting" />
 
     <fragment
         android:id="@+id/nav_share"
