+++ app/src/main/res/values/strings.xml
@@ -14,7 +14,7 @@
     <string name="menu_home">Home</string>
     <string name="menu_gallery">Gallery</string>
     <string name="menu_video">Videos</string>
-    <string name="menu_tools">Tools</string>
+    <string name="menu_setting">Setting</string>
     <string name="menu_share">Share</string>
     <string name="menu_send">Send</string>
     <string name="menu_hotels">Hotels</string>
